net: dovetail: drop SCH_OOB queuing discipline
We could not efficiently provide DMA pre-mapping for the buffers
flowing through this qdisc in order to handle them out-of-band without
further requirement for DMA streaming operations in the NIC driver.

Besides, there is no practical upside of this queuing discipline over
allowing NIC drivers to handle concurrent in-band and out-of-band
traffic to flow through them, which is achievable via proper use of
stage exclusion locks.

IOW, we don't need that qdisc.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>