evl/net: dev: enable real devices as oob ports
Since we may use an eBPF program to discriminate the oob traffic
inside the input flow in replacement for VLAN-based selection, we
may now allow real netdevs to be used as oob ports too.

The egress path was already covered by the work on the UDP stack, with
the VLAN tagging only applied to the outgoing packets when applicable
(i.e. when the output device to the next hop is indeed a VLAN
interface).

Signed-off-by: Philippe Gerum <rpm@xenomai.org>